Needing to almost restart project for Azure AD B2C
- Apparently if I want a setup where any user can login with SSO on my
website I really need to use Azure AD B2C which is designed for that
kind of purpose. However, unfortunately the configuration settings
for each seem to be their own separate thing which means I'll
likely have to find a different guide/tutorial to help me setup the
configuration for the B2C version rather than the normal version.
- Note for clarity. Azure AD has been renamed to Microsoft Entra Id.
So any place where I mention Azure AD in the future (without the B2C)
I'm referring to the new "Microsoft Entra ID". It's just simply been
renamed.
- I think I will likely restart the entire project with new
project generations for the front-end and the backend and try to
find a matching tutorial for the B2C version.